test: make -Wattribute-warning a warning in unavailable test
Basically, HEDLEY_UNAVAILABLE will use the warning attribute to get the
compiler to emit a warning when the symbol is used, which is exactly
what we want. However, recent versions of clang will turn that warning
into an error if -Werror is specified. This is a great feature, but
it causes our test to fail when we don't want it to.