ci: expand NVHPC coverage
There are really two parts here: using earlier versions of Ubuntu to
test older versions of the compilers, and trying to figure out which
versions could be run on 24.04 instead of 22.04 (current "latest").
I'd like to use the newest distro I can since most CI providers only
support the few most recent LTS releases. Building on the newest
distro which supports the compiler is likely to give us the longest
window of support.